削除ボタン後にダイアログで削除決定させる処理を追加

+++ apps/frontend/my-app/src/Component/Contentinfo.js
@@ -11,13 +11,19 @@ function Contentinfo({ src, alt, username, userid, postid, content, imagepath, h
   const currentUserId = user?.id;
 
   const handleClick = async () => {
+    const confirmDelete = window.confirm("本当にこのポストを削除しますか？");
+    if (!confirmDelete) {
+      return; // キャンセルされた場合は何もしない
+    }
+
+
     try {
       await deletePost(postid, userid);
       handleDelete(postid);
       alert('ポストが削除されました');
     } catch (error) {
       alert('ポストの削除に失敗しました');
-    }
+    } 
   };
 
   return (
